Add tests for hook validation
